Improve configuration exception handling
Since moving version out of the load_spec and into options, we were
producing a false positive in load(): successfully generating a load_spec
that specified the webfont method, even when the options included
both version 5.0.9 and v4shim. But an exception was being thrown
on enqueue() because the release provider knows that we don't have
a v4shim for the webfont method before 5.1.0.

This shows that we need to do a better job of handling the scope of
possible configuration conflicts, just within the computation of the
load_spec, but across the combined input of load_spec + options.

We've needed to add a mechanism for bubbling up exceptions with
user-facing notices anyway, so this is a good initial case to try that.