Only allow removal of API Token when no kit is selected
If a kit is active or even pending, the "remove" link disappears.
This may not be the final solution, but it seems like the easiest
way for now to keep the user from putting himself into an inconsistent
state.

The problem with it might be that it could be confusing for the user
to figure out what he needs to do to get the remove link to show up.
The answer: switch to CDN config mode, make that active, and then
switch over to the kits settings view just to click "remove" on
  the API Token. Not a great "remove API Token" UX.