Rename pseudoElements option and revise its semantics
It will be true whenenver Font Awesome is configured in a way
that supports CSS Pseudo-elements.

So when configured for webfont technology, the configuration is only
valid when pseudoElements is true.

The front-end code will coerce and default the pseudoElements option
appropriately as technology changes, both in the CDN and Kits
settings tabs.